requirements.txt
openai
python-dotenv
gradio
matplotlib
google-generativeai
Pillow
